refactor: remove browser-specific code in cdp
