+++ packages/puppeteer-core/src/cdp/CDPSession.ts
@@ -76,7 +76,8 @@ export class CdpCDPSession extends CDPSession {
 
   override parentSession(): CDPSession | undefined {
     if (!this.#parentSessionId) {
-      // To make it work in Firefox that does not have parent (tab) sessions.
+      // In some cases, e.g., DevTools pages there is no parent session. In this
+      // case, we treat the current session as the parent session.
       return this;
     }
     const parent = this.#connection?.session(this.#parentSessionId);
