+++ packages/puppeteer-core/src/cdp/ExecutionContext.ts
@@ -235,9 +235,8 @@ export class ExecutionContext
     try {
       await this.#addBinding(binding);
     } catch (err) {
-      // If the binding cannot be added, then either the browser doesn't support
-      // bindings (e.g. Firefox) or the context is broken. Either breakage is
-      // okay, so we ignore the error.
+      // If the binding cannot be added, the context is broken. We cannot
+      // recover so we ignore the error.
       debugError(err);
     }
   }
