+++ packages/puppeteer-core/src/cdp/Page.ts
@@ -73,8 +73,8 @@ import {MAIN_WORLD} from './IsolatedWorlds.js';
 import {releaseObject} from './JSHandle.js';
 import type {NetworkConditions} from './NetworkManager.js';
 import type {CdpTarget} from './Target.js';
+import {TargetManagerEvent} from './TargetManageEvents.js';
 import type {TargetManager} from './TargetManager.js';
-import {TargetManagerEvent} from './TargetManager.js';
 import {Tracing} from './Tracing.js';
 import {
   createClientError,
@@ -1004,7 +1004,6 @@ export class CdpPage extends Page {
     } = options;
 
     await using stack = new AsyncDisposableStack();
-    // Firefox omits background by default; it's not configurable.
     if (omitBackground && (type === 'png' || type === 'webp')) {
       await this.#emulationManager.setTransparentBackgroundColor();
       stack.defer(async () => {
@@ -1030,15 +1029,14 @@ export class CdpPage extends Page {
       clip = getIntersectionRect(clip, viewport);
     }
 
-    // We need to do these spreads because Firefox doesn't allow unknown options.
     const {data} = await this.#primaryTargetClient.send(
       'Page.captureScreenshot',
       {
         format: type,
-        ...(optimizeForSpeed ? {optimizeForSpeed} : {}),
+        optimizeForSpeed,
+        fromSurface,
         ...(quality !== undefined ? {quality: Math.round(quality)} : {}),
         ...(clip ? {clip: {...clip, scale: clip.scale ?? 1}} : {}),
-        ...(!fromSurface ? {fromSurface} : {}),
         captureBeyondViewport,
       },
     );
